cleanup wallet files
