+++ src-tauri/src/main.rs
@@ -51,7 +51,6 @@ use tokio::sync::watch::{self};
 use tor_control_client::TorStatus;
 use updates_manager::UpdatesManager;
 use utils::system_status::SystemStatus;
-use wallet_adapter::WalletState;
 use websocket_events_manager::WebsocketEventsManager;
 use websocket_manager::{WebsocketManager, WebsocketManagerStatusMessage, WebsocketMessage};
 
@@ -81,7 +80,8 @@ use crate::node::node_manager::NodeManager;
 use crate::p2pool::models::P2poolStats;
 use crate::p2pool_manager::P2poolManager;
 use crate::tor_manager::TorManager;
-use crate::wallet_manager::WalletManager;
+use crate::wallet::wallet_manager::WalletManager;
+use crate::wallet::wallet_types::WalletState;
 
 mod ab_test_selector;
 mod airdrop;
@@ -128,7 +128,6 @@ mod progress_trackers;
 mod release_notes;
 mod requests;
 mod setup;
-mod spend_wallet;
 mod systemtray_manager;
 mod tapplets;
 mod tasks_tracker;
@@ -140,8 +139,7 @@ mod tor_control_client;
 mod tor_manager;
 mod updates_manager;
 mod utils;
-mod wallet_adapter;
-mod wallet_manager;
+mod wallet;
 mod websocket_events_manager;
 mod websocket_manager;
 mod xmrig;
