+++ src-tauri/src/wallet/mod.rs
@@ -20,20 +20,9 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-pub mod address_utils;
-pub mod app_flow_utils;
-pub mod commands_builder;
-pub mod cryptography;
-pub mod file_utils;
-pub mod formatting_utils;
-pub mod locks_utils;
-pub mod logging_utils;
-pub mod macos_utils;
-pub mod math_utils;
-pub mod network_status;
-pub mod platform_utils;
-pub mod rand_utils;
-
-pub mod system_status;
-#[cfg(windows)]
-pub mod windows_setup_utils;
+pub mod spend_wallet;
+pub mod transaction_service;
+pub mod wallet_adapter;
+pub mod wallet_manager;
+pub mod wallet_status_monitor;
+pub mod wallet_types;
